detect-engine-mpm: change transforms->cnt to uint8_t

+++ src/detect-engine.h
@@ -61,7 +61,7 @@ bool DetectEngineBufferTypeSupportsMultiInstanceGetById(
 bool DetectEngineBufferTypeSupportsFramesGetById(const DetectEngineCtx *de_ctx, const int id);
 const char *DetectEngineBufferTypeGetDescriptionById(const DetectEngineCtx *de_ctx, const int id);
 int DetectEngineBufferTypeGetByIdTransforms(
-        DetectEngineCtx *de_ctx, const int id, TransformData *transforms, int transform_cnt);
+        DetectEngineCtx *de_ctx, const int id, TransformData *transforms, uint8_t transform_cnt);
 void DetectEngineBufferRunSetupCallback(const DetectEngineCtx *de_ctx, const int id, Signature *s);
 bool DetectEngineBufferRunValidateCallback(
         const DetectEngineCtx *de_ctx, const int id, const Signature *s, const char **sigerror);
